move FetchResponse types to public_types
