+++ packages/remote-config/src/public_types.ts
@@ -16,29 +16,6 @@
  */
 
 import { FirebaseApp } from '@firebase/app';
-import { FetchResponse } from './client/remote_config_fetch_client';
-
-export {
-  FetchResponse,
-  FirebaseRemoteConfigObject
-} from './client/remote_config_fetch_client';
-
-/**
- * Options for Remote Config initialization.
- *
- * @public
- */
-export interface RemoteConfigOptions {
-  /**
-   * The ID of the template to use. If not provided, defaults to "firebase".
-   */
-  templateId?: string;
-
-  /**
-   * Hydrates the state with an initial fetch response.
-   */
-  initialFetchResponse?: FetchResponse;
-}
 
 /**
  * The Firebase Remote Config service interface.
@@ -71,6 +48,67 @@ export interface RemoteConfig {
    * The status of the last fetch <i>attempt</i>.
    */
   lastFetchStatus: FetchStatus;
+}
+
+/**
+ * Defines a self-descriptive reference for config key-value pairs.
+ */
+export interface FirebaseRemoteConfigObject {
+  [key: string]: string;
+}
+
+/**
+ * Defines a successful response (200 or 304).
+ *
+ * <p>Modeled after the native {@link Response} interface, but simplified for Remote Config's
+ * use case.
+ */
+export interface FetchResponse {
+  /**
+   * The HTTP status, which is useful for differentiating success responses with data from
+   * those without.
+   *
+   * <p>{@link RemoteConfigClient} is modeled after the native {@link GlobalFetch} interface, so
+   * HTTP status is first-class.
+   *
+   * <p>Disambiguation: the fetch response returns a legacy "state" value that is redundant with the
+   * HTTP status code. The former is normalized into the latter.
+   */
+  status: number;
+
+  /**
+   * Defines the ETag response header value.
+   *
+   * <p>Only defined for 200 and 304 responses.
+   */
+  eTag?: string;
+
+  /**
+   * Defines the map of parameters returned as "entries" in the fetch response body.
+   *
+   * <p>Only defined for 200 responses.
+   */
+  config?: FirebaseRemoteConfigObject;
+
+  // Note: we're not extracting experiment metadata until
+  // ABT and Analytics have Web SDKs.
+}
+
+/**
+ * Options for Remote Config initialization.
+ *
+ * @public
+ */
+export interface RemoteConfigOptions {
+  /**
+   * The ID of the template to use. If not provided, defaults to "firebase".
+   */
+  templateId?: string;
+
+  /**
+   * Hydrates the state with an initial fetch response.
+   */
+  initialFetchResponse?: FetchResponse;
 }
 
 /**
