+++ packages/remote-config/src/storage/storage_cache.ts
@@ -16,14 +16,14 @@
  */
 
 import { FetchStatus, CustomSignals } from '@firebase/remote-config-types';
-import { FirebaseRemoteConfigObject } from '../client/remote_config_fetch_client';
+import { FirebaseRemoteConfigObject } from '../public_types';
 import { Storage } from './storage';
 
 /**
  * A memory cache layer over storage to support the SDK's synchronous read requirements.
  */
 export class StorageCache {
-  constructor(private readonly storage: Storage) {}
+  constructor(private readonly storage: Storage) { }
 
   /**
    * Memory caches.
