+++ packages/remote-config/test/remote_config.test.ts
@@ -17,6 +17,7 @@
 
 import { FirebaseApp } from '@firebase/app';
 import {
+  FetchResponse,
   RemoteConfig as RemoteConfigType,
   LogLevel as RemoteConfigLogLevel
 } from '../src/public_types';
@@ -27,7 +28,6 @@ import { Storage } from '../src/storage/storage';
 import { RemoteConfig } from '../src/remote_config';
 import {
   RemoteConfigFetchClient,
-  FetchResponse
 } from '../src/client/remote_config_fetch_client';
 import { Value } from '../src/value';
 import './setup';
